% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WT.R
\name{clearVar}
\alias{clearVar}
\title{Clears all variables (and functions)}
\usage{
clearVar(env)
}
\arguments{
\item{env}{specific environment to apply to.  Default: .GlobalEnv}
}
\description{
Clears all variables (and functions)
}
