% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WT.R
\name{colToDate}
\alias{colToDate}
\title{Converts all data in column to Date datatype}
\usage{
colToDate(df, colName, date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{original dataframe}

\item{colName}{column to convert}

\item{date_format}{the format the string is in. By default, ISO standard}
}
\value{
dataframe
}
\description{
Converts all data in column to Date datatype
}
