% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WT.R
\name{colToNum}
\alias{colToNum}
\title{Converts all data in column to numeric (float) datatype}
\usage{
colToNum(df, colName)
}
\arguments{
\item{df}{original dataframe}

\item{colName}{column to convert}
}
\value{
dataframe
}
\description{
Converts all data in column to numeric (float) datatype
}
