% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WT.R
\name{freadc}
\alias{freadc}
\title{Reads df from clipboard}
\usage{
freadc(...)
}
\arguments{
\item{...}{standard fread arguments}
}
\value{
dataframe
}
\description{
Takes copied data from clipboard memory, saves to file where it is then loaded with fread
}
